requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
jinja2>=3.1.2
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
requests>=2.31.0
beautifulsoup4>=4.12.3
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
sqlalchemy-utils>=0.41.1
